Refactor GUID normalization into shared utility
