+++ swifttunnel-core/src/settings.rs
@@ -5,6 +5,7 @@
 use crate::network_analyzer::NetworkTestResultsCache;
 use crate::structs::Config;
 use crate::updater::{UpdateChannel, UpdateSettings};
+use crate::utils::normalize_guid_ascii_lowercase;
 use log::{debug, error, info};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -201,63 +202,6 @@ fn default_server() -> String {
 
 fn default_game_presets() -> Vec<String> {
     vec!["roblox".to_string()] // Default to Roblox selected
-}
-
-fn normalize_guid_ascii_lowercase(value: &str) -> Option<String> {
-    fn is_guid_ascii(bytes: &[u8]) -> bool {
-        if bytes.len() != 36 {
-            return false;
-        }
-        const DASH_POS: [usize; 4] = [8, 13, 18, 23];
-        for (i, &b) in bytes.iter().enumerate() {
-            if DASH_POS.contains(&i) {
-                if b != b'-' {
-                    return false;
-                }
-                continue;
-            }
-            if !b.is_ascii_hexdigit() {
-                return false;
-            }
-        }
-        true
-    }
-
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    let bytes = trimmed.as_bytes();
-
-    // Fast path: `{GUID}` anywhere inside the string.
-    for (open_idx, &b) in bytes.iter().enumerate() {
-        if b != b'{' {
-            continue;
-        }
-        let Some(close_rel) = bytes[open_idx + 1..].iter().position(|&b| b == b'}') else {
-            continue;
-        };
-        let close_idx = open_idx + 1 + close_rel;
-        let inner = &bytes[open_idx + 1..close_idx];
-        if is_guid_ascii(inner) {
-            return trimmed
-                .get(open_idx + 1..close_idx)
-                .map(|guid| guid.to_ascii_lowercase());
-        }
-    }
-
-    // Fallback: raw GUID without braces somewhere in the string.
-    for start in 0..=bytes.len().saturating_sub(36) {
-        let candidate = &bytes[start..start + 36];
-        if is_guid_ascii(candidate) {
-            return trimmed
-                .get(start..start + 36)
-                .map(|guid| guid.to_ascii_lowercase());
-        }
-    }
-
-    None
 }
 
 impl Default for AppSettings {
